moved `staff` package into `commands` package
