+++ src/main/java/net/javadiscord/javabot/systems/commands/staff/EditEmbedCommand.java
@@ -1,4 +1,4 @@
-package net.javadiscord.javabot.systems.staff;
+package net.javadiscord.javabot.systems.commands.staff;
 
 import net.dv8tion.jda.api.EmbedBuilder;
 import net.dv8tion.jda.api.entities.Message;
