+++ src/main/java/net/javadiscord/javabot/systems/commands/staff/custom_commands/subcommands/CustomCommandCreate.java
@@ -1,6 +1,5 @@
-package net.javadiscord.javabot.systems.staff.custom_commands.subcommands;
+package net.javadiscord.javabot.systems.commands.staff.custom_commands.subcommands;
 
-import com.mongodb.BasicDBObject;
 import com.mongodb.client.MongoCollection;
 import net.dv8tion.jda.api.EmbedBuilder;
 import net.dv8tion.jda.api.events.interaction.SlashCommandEvent;
@@ -10,41 +9,56 @@ import net.javadiscord.javabot.Bot;
 import net.javadiscord.javabot.command.Responses;
 import net.javadiscord.javabot.command.SlashCommandHandler;
 import net.javadiscord.javabot.events.StartupListener;
-import net.javadiscord.javabot.systems.staff.custom_commands.CustomCommandHandler;
+import net.javadiscord.javabot.systems.commands.staff.custom_commands.CustomCommandHandler;
 import org.bson.Document;
 
 import java.time.Instant;
 
 /**
- * Subcommand that allows to delete Custom Slash Commands. {@link CustomCommandHandler#CustomCommandHandler()}
+ * Subcommand that allows to create Custom Slash Commands. {@link CustomCommandHandler#CustomCommandHandler()}
  */
-public class CustomCommandDelete implements SlashCommandHandler {
+public class CustomCommandCreate implements SlashCommandHandler {
     @Override
     public ReplyAction handle(SlashCommandEvent event) {
-        OptionMapping nameOption = event.getOption("name");
+        OptionMapping replyOption = event.getOption("reply");
+        boolean reply = replyOption == null || replyOption.getAsBoolean();
 
-        if (nameOption == null) {
+        OptionMapping embedOption = event.getOption("embed");
+        boolean embed = embedOption == null || embedOption.getAsBoolean();
+
+        OptionMapping nameOption = event.getOption("name");
+        OptionMapping textOption = event.getOption("text");
+
+        if (nameOption == null || textOption == null) {
             return Responses.error(event, "Missing required arguments.");
         }
+
         String name = nameOption.getAsString();
+        String text = textOption.getAsString();
 
         MongoCollection<Document> collection = StartupListener.mongoClient
                 .getDatabase("other")
                 .getCollection("customcommands");
 
-        if (!CustomCommandHandler.commandExists(event.getGuild().getId(), name)) {
-            return Responses.error(event, "A Custom Slash Command called `" + "/" + name + "` does not exist.");
+        if (CustomCommandHandler.commandExists(event.getGuild().getId(), name)) {
+            return Responses.error(event, "A Custom Slash Command called " + "`" + "/" + name + "` already exists.");
         }
 
-       collection.deleteOne(
-                new BasicDBObject()
+        collection.insertOne(
+                new Document()
                 .append("guildId", event.getGuild().getId())
                 .append("commandName", name)
+                .append("value", text)
+                .append("reply", reply)
+                .append("embed", embed)
         );
 
         var e = new EmbedBuilder()
-                .setTitle("Custom Slash Command deleted")
+                .setTitle("Custom Command created")
                 .addField("Name", "```" + "/" + name + "```", false)
+                .addField("Value", "```" + text + "```", false)
+                .addField("Reply?", "`" + reply + "`", true)
+                .addField("Embed?", "`" + embed + "`", true)
                 .setFooter(event.getUser().getAsTag(), event.getUser().getEffectiveAvatarUrl())
                 .setColor(Bot.config.get(event.getGuild()).getSlashCommand().getDefaultColor())
                 .setTimestamp(Instant.now())
