+++ src/main/java/net/javadiscord/javabot/systems/commands/staff/custom_commands/subcommands/CustomCommandDelete.java
@@ -1,4 +1,4 @@
-package net.javadiscord.javabot.systems.staff.custom_commands.subcommands;
+package net.javadiscord.javabot.systems.commands.staff.custom_commands.subcommands;
 
 import com.mongodb.BasicDBObject;
 import com.mongodb.client.MongoCollection;
@@ -10,32 +10,23 @@ import net.javadiscord.javabot.Bot;
 import net.javadiscord.javabot.command.Responses;
 import net.javadiscord.javabot.command.SlashCommandHandler;
 import net.javadiscord.javabot.events.StartupListener;
-import net.javadiscord.javabot.systems.staff.custom_commands.CustomCommandHandler;
+import net.javadiscord.javabot.systems.commands.staff.custom_commands.CustomCommandHandler;
 import org.bson.Document;
 
 import java.time.Instant;
 
 /**
- * Subcommand that allows to edit Custom Slash Commands. {@link CustomCommandHandler#CustomCommandHandler()}
+ * Subcommand that allows to delete Custom Slash Commands. {@link CustomCommandHandler#CustomCommandHandler()}
  */
-public class CustomCommandEdit implements SlashCommandHandler {
+public class CustomCommandDelete implements SlashCommandHandler {
     @Override
     public ReplyAction handle(SlashCommandEvent event) {
-        OptionMapping replyOption = event.getOption("reply");
-        boolean reply = replyOption == null || replyOption.getAsBoolean();
-
-        OptionMapping embedOption = event.getOption("embed");
-        boolean embed = embedOption == null || embedOption.getAsBoolean();
-
         OptionMapping nameOption = event.getOption("name");
-        OptionMapping textOption = event.getOption("text");
 
-        if (nameOption == null || textOption == null) {
+        if (nameOption == null) {
             return Responses.error(event, "Missing required arguments.");
         }
-
         String name = nameOption.getAsString();
-        String text = textOption.getAsString();
 
         MongoCollection<Document> collection = StartupListener.mongoClient
                 .getDatabase("other")
@@ -45,23 +36,15 @@ public class CustomCommandEdit implements SlashCommandHandler {
             return Responses.error(event, "A Custom Slash Command called `" + "/" + name + "` does not exist.");
         }
 
-        collection.updateOne(
+       collection.deleteOne(
                 new BasicDBObject()
                 .append("guildId", event.getGuild().getId())
-                .append("commandName", name),
-                new Document("$set",
-                        new Document("value", text)
-                                .append("reply", reply)
-                                .append("embed", embed)
-                )
+                .append("commandName", name)
         );
 
         var e = new EmbedBuilder()
-                .setTitle("Custom Slash Command edited")
+                .setTitle("Custom Slash Command deleted")
                 .addField("Name", "```" + "/" + name + "```", false)
-                .addField("Value", "```" + text + "```", false)
-                .addField("Reply?", "`" + reply + "`", true)
-                .addField("Embed?", "`" + embed + "`", true)
                 .setFooter(event.getUser().getAsTag(), event.getUser().getEffectiveAvatarUrl())
                 .setColor(Bot.config.get(event.getGuild()).getSlashCommand().getDefaultColor())
                 .setTimestamp(Instant.now())
