+++ src/main/java/net/javadiscord/javabot/systems/commands/staff/custom_commands/subcommands/CustomCommandEdit.java
@@ -1,5 +1,6 @@
-package net.javadiscord.javabot.systems.staff.custom_commands.subcommands;
+package net.javadiscord.javabot.systems.commands.staff.custom_commands.subcommands;
 
+import com.mongodb.BasicDBObject;
 import com.mongodb.client.MongoCollection;
 import net.dv8tion.jda.api.EmbedBuilder;
 import net.dv8tion.jda.api.events.interaction.SlashCommandEvent;
@@ -9,15 +10,15 @@ import net.javadiscord.javabot.Bot;
 import net.javadiscord.javabot.command.Responses;
 import net.javadiscord.javabot.command.SlashCommandHandler;
 import net.javadiscord.javabot.events.StartupListener;
-import net.javadiscord.javabot.systems.staff.custom_commands.CustomCommandHandler;
+import net.javadiscord.javabot.systems.commands.staff.custom_commands.CustomCommandHandler;
 import org.bson.Document;
 
 import java.time.Instant;
 
 /**
- * Subcommand that allows to create Custom Slash Commands. {@link CustomCommandHandler#CustomCommandHandler()}
+ * Subcommand that allows to edit Custom Slash Commands. {@link CustomCommandHandler#CustomCommandHandler()}
  */
-public class CustomCommandCreate implements SlashCommandHandler {
+public class CustomCommandEdit implements SlashCommandHandler {
     @Override
     public ReplyAction handle(SlashCommandEvent event) {
         OptionMapping replyOption = event.getOption("reply");
@@ -40,21 +41,23 @@ public class CustomCommandCreate implements SlashCommandHandler {
                 .getDatabase("other")
                 .getCollection("customcommands");
 
-        if (CustomCommandHandler.commandExists(event.getGuild().getId(), name)) {
-            return Responses.error(event, "A Custom Slash Command called " + "`" + "/" + name + "` already exists.");
+        if (!CustomCommandHandler.commandExists(event.getGuild().getId(), name)) {
+            return Responses.error(event, "A Custom Slash Command called `" + "/" + name + "` does not exist.");
         }
 
-        collection.insertOne(
-                new Document()
+        collection.updateOne(
+                new BasicDBObject()
                 .append("guildId", event.getGuild().getId())
-                .append("commandName", name)
-                .append("value", text)
-                .append("reply", reply)
-                .append("embed", embed)
+                .append("commandName", name),
+                new Document("$set",
+                        new Document("value", text)
+                                .append("reply", reply)
+                                .append("embed", embed)
+                )
         );
 
         var e = new EmbedBuilder()
-                .setTitle("Custom Command created")
+                .setTitle("Custom Slash Command edited")
                 .addField("Name", "```" + "/" + name + "```", false)
                 .addField("Value", "```" + text + "```", false)
                 .addField("Reply?", "`" + reply + "`", true)
