+++ src/main/java/net/javadiscord/javabot/systems/commands/staff/reaction_roles/ReactionRoleCommandHandler.java
@@ -1,11 +1,11 @@
-package net.javadiscord.javabot.systems.staff.reaction_roles;
+package net.javadiscord.javabot.systems.commands.staff.reaction_roles;
 
 import net.dv8tion.jda.api.events.interaction.SlashCommandEvent;
 import net.dv8tion.jda.api.requests.restaction.interactions.ReplyAction;
 import net.javadiscord.javabot.command.DelegatingCommandHandler;
 import net.javadiscord.javabot.command.Responses;
-import net.javadiscord.javabot.systems.staff.reaction_roles.subcommands.CreateReactionRole;
-import net.javadiscord.javabot.systems.staff.reaction_roles.subcommands.DeleteReactionRole;
+import net.javadiscord.javabot.systems.commands.staff.reaction_roles.subcommands.CreateReactionRole;
+import net.javadiscord.javabot.systems.commands.staff.reaction_roles.subcommands.DeleteReactionRole;
 
 public class ReactionRoleCommandHandler extends DelegatingCommandHandler {
